refactor: split three files that crossed the 500-LoC cap as a consequence of stacking
Stacking Wave 1 pushed three files over the cap. Each is attributable to one
PR, and none of the PR branches amended the allow-list — so each would be red
on the cap check standalone:

| file | before → after | from | extracted to |
|---|---|---|---|
| `spark-runtime/src/weights.rs` | 458 → 506 → **412** | #439 | `weights/teardown_tests.rs` |
| `spark-server/src/api/tests/sanitizer.rs` | 199 → 502 → **205** | #206 | `sanitizer_chunk_tests.rs` |
| `spark-model/.../trait_impl/decode_a.rs` | 500 → 518 → **420** | #372 | `decode_a3.rs` |

## Why split rather than allow-list

`file-size-cap.yml` states that adding an entry requires a rationale comment
**and a tracking issue**. That issue belongs to the PR author, not to whoever
happens to be assembling the stack — and quietly widening a cap on a
contributor's behalf is the wrong default. The workflow's own error text names
splitting as the remedy, so that is what this does.

Each cut follows the neighbouring convention rather than inventing a layout:
`sanitizer_chunk_tests.rs` sits beside the existing `envelope.rs` /
`harness.rs` / `watchdog.rs` test siblings, and `decode_a3.rs` beside the
existing `decode_a2.rs`.

**Re-organisation only — no behaviour change.** These are someone else's PRs;
the test count is unchanged at 3998 passed / 0 failed, which is the check that
matters: a drop would have meant a test was moved out of the build rather than
into a sibling.

clippy + fmt clean; 0 LoC violations against the workflow's own 94-entry
allow-list.
